Refactor code, edit naming convention
