+++ hotel-management/package.json
@@ -23,6 +23,8 @@
     ]
   },
   "dependencies": {
+    "@types/prop-types": "^15.7.9",
+    "prop-types": "^15.8.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-icons": "^4.11.0",
