+++ hotel-management/src/components/AppLayout.tsx
