+++ hotel-management/src/components/Empty.tsx
@@ -1,13 +1,13 @@
 import styled from 'styled-components';
 
-const StyledEmpty = styled.p`
+const StyledMessage = styled.p`
   font-size: var(--fs-sm);
   text-align: center;
   padding: 20px;
 `;
 
 const Empty = ({ children }: { children: string }) => {
-  return <StyledEmpty>{children}</StyledEmpty>;
+  return <StyledMessage>{children}</StyledMessage>;
 };
 
 export default Empty;
