+++ hotel-management/src/components/Nav.tsx
