+++ hotel-management/src/components/Sidebar.tsx
