+++ hotel-management/src/components/Table.tsx
