+++ hotel-management/src/globals/types.ts
@@ -1,0 +1,8 @@
+export type TUser = {
+  id: string;
+  name: string;
+  identifiedCode: string;
+  address: string;
+  phone: string;
+  roomId: string;
+};
